requirements.txt
asgiref==3.5.2
Django==4.0.5
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-taggit==3.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
